refactor: Rename participant model filename
