+++ lib/src/auth/participant.dart
