+++ test/src/auth/auth_test.dart
@@ -1,6 +1,6 @@
 import 'package:firebase_auth_mocks/firebase_auth_mocks.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:mdigit_span_tasks_ema/src/services/auth.dart';
+import 'package:mdigit_span_tasks_ema/src/auth/auth.dart';
 
 void main() {
   test(
